requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0        
tensorflow==2.17.0
keras==3.6.0
numpy==1.24.3
Pillow==10.4.0           
opencv-python==4.10.0.84
pymongo==4.10.1
python-dotenv==1.0.1
geocoder>=1.38.1
requests>=2.31.0